+++ src/main/java/com/genius/tms_c61_genius/service/ServiceImpl/GenreServiceImpl.java
@@ -1,5 +1,7 @@
 package com.genius.tms_c61_genius.service.ServiceImpl;
 
+import com.genius.tms_c61_genius.exception.BadDataException;
+import com.genius.tms_c61_genius.exception.NotFoundException;
 import com.genius.tms_c61_genius.model.domain.Genre;
 import com.genius.tms_c61_genius.repository.GenreRepository;
 import com.genius.tms_c61_genius.service.GenreService;
@@ -16,6 +18,8 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public String createGenre(String genreName) {
+        if(genreRepository.existsGenreByGenreTitle(genreName))
+            throw new BadDataException("genre with such title is already exist");
         return genreRepository.save(Genre.builder()
                 .genreTitle(genreName)
                 .build()).getGenreTitle();
@@ -23,6 +27,10 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public String updateGenre(Integer genreId, String genreName) {
+        if(genreRepository.existsGenreByGenreTitle(genreName))
+            throw new BadDataException("genre with such title is already exist");
+        if(!genreRepository.existsGenreById(genreId))
+            throw new NotFoundException("genre not found");
         Genre updatedGenre = genreRepository.getGenreById(genreId);
         updatedGenre.setGenreTitle(genreName);
         return genreRepository.save(updatedGenre).getGenreTitle();
@@ -30,11 +38,15 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public void deleteGenre(Integer genreId) {
+        if(!genreRepository.existsGenreById(genreId))
+            throw new NotFoundException("genre not found");
         genreRepository.deleteById(genreId);
     }
 
     @Override
     public String getGenre(Integer genreId) {
+        if(!genreRepository.existsGenreById(genreId))
+            throw new NotFoundException("genre not found");
         return genreRepository.getGenreById(genreId).getGenreTitle();
     }
 }
