+++ src/main/java/com/genius/tms_c61_genius/service/ServiceImpl/LabelServiceImpl.java
@@ -1,5 +1,7 @@
 package com.genius.tms_c61_genius.service.ServiceImpl;
 
+import com.genius.tms_c61_genius.exception.BadDataException;
+import com.genius.tms_c61_genius.exception.NotFoundException;
 import com.genius.tms_c61_genius.mapper.LabelDtoMapper;
 import com.genius.tms_c61_genius.model.domain.Label;
 import com.genius.tms_c61_genius.model.request.LabelReqDto;
@@ -21,18 +23,23 @@ public class LabelServiceImpl implements LabelService {
 
     @Override
     public LabelResDto createLabel(LabelReqDto labelReqDto) {
-        System.out.println(labelReqDto);
+        if(labelRepository.existsLabelByLabelName(labelReqDto.getLabelName()))
+            throw new BadDataException("label with such name is already exist");
         Label savedLabel = labelRepository.save(labelDtoMapper.labelReqToLabel(labelReqDto));
         return labelDtoMapper.labelToLabelRes(savedLabel);
     }
 
     @Override
     public LabelResDto getLabelInfo(String labelName) {
+        if(!labelRepository.existsLabelByLabelName(labelName))
+            throw new NotFoundException("label not found");
         return labelDtoMapper.labelToLabelRes(labelRepository.getLabelByLabelName(labelName));
     }
 
     @Override
     public LabelResDto updateLabelInfo(LabelReqDto labelReqDto) {
+        if(!labelRepository.existsLabelByLabelName(labelReqDto.getLabelName()))
+            throw new NotFoundException("label not found");
         Label updatedLabel = labelRepository.getLabelByLabelName(labelReqDto.getLabelName());
         updatedLabel.setDescription(labelReqDto.getDescription());
         return labelDtoMapper.labelToLabelRes(updatedLabel);
@@ -40,6 +47,8 @@ public class LabelServiceImpl implements LabelService {
 
     @Override
     public void deleteLabel(String labelName) {
+        if(!labelRepository.existsLabelByLabelName(labelName))
+            throw new NotFoundException("label not found");
         labelRepository.delete(labelRepository.getLabelByLabelName(labelName));
     }
 }
