+++ src/main/java/com/genius/tms_c61_genius/service/ServiceImpl/RoleServiceImpl.java
@@ -1,5 +1,7 @@
 package com.genius.tms_c61_genius.service.ServiceImpl;
 
+import com.genius.tms_c61_genius.exception.BadDataException;
+import com.genius.tms_c61_genius.exception.NotFoundException;
 import com.genius.tms_c61_genius.model.domain.Role;
 import com.genius.tms_c61_genius.repository.RoleRepository;
 import com.genius.tms_c61_genius.service.RoleService;
@@ -16,6 +18,8 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public String createRole(String roleName) {
+        if(roleRepository.existsRoleByRoleName(roleName))
+            throw new BadDataException("role with such name is already exist");
         return roleRepository.save(Role.builder()
                 .roleName(roleName)
                 .build()).getRoleName();
@@ -23,6 +27,8 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public String updateRole(Integer roleId,String roleName) {
+        if(!roleRepository.existsRoleById(roleId))
+            throw new NotFoundException("role not found");
         Role updatedRole = roleRepository.getRoleById(roleId);
         updatedRole.setRoleName(roleName);
         return roleRepository.save(updatedRole).getRoleName();
@@ -30,11 +36,15 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public void deleteRole(Integer roleId) {
+        if(!roleRepository.existsRoleById(roleId))
+            throw new NotFoundException("role not found");
         roleRepository.deleteById(roleId);
     }
 
     @Override
     public String getRole(Integer roleId) {
+        if(!roleRepository.existsRoleById(roleId))
+            throw new NotFoundException("role not found");
         return roleRepository.getRoleById(roleId).getRoleName();
     }
 }
