+++ src/main/java/com/genius/tms_c61_genius/service/ServiceImpl/SoundProducerServiceImpl.java
@@ -1,5 +1,7 @@
 package com.genius.tms_c61_genius.service.ServiceImpl;
 
+import com.genius.tms_c61_genius.exception.BadDataException;
+import com.genius.tms_c61_genius.exception.NotFoundException;
 import com.genius.tms_c61_genius.mapper.SoundProducerDtoMapper;
 import com.genius.tms_c61_genius.model.domain.Artist;
 import com.genius.tms_c61_genius.model.domain.SoundProducer;
@@ -24,6 +26,8 @@ public class SoundProducerServiceImpl implements SoundProducerService {
     @Override
     @Transactional
     public SoundProducerResDto createSoundProducer(SoundProducerReqDto soundProducerReqDto) {
+        if(soundProducerRepository.existsSoundProducerByPersonInfoNickname(soundProducerReqDto.getNickname()))
+            throw new BadDataException("producer with such nickname is already exist");
         SoundProducer savedProducer =  soundProducerDtoMapper.producerReqToProducer(soundProducerReqDto);
         soundProducerRepository.save(savedProducer);
         return soundProducerDtoMapper.producerToProducerResDto(savedProducer);
@@ -32,18 +36,24 @@ public class SoundProducerServiceImpl implements SoundProducerService {
     @Override
     @Transactional
     public void deleteProducerById(Integer id) {
+        if(!soundProducerRepository.existsSoundProducerById(id))
+            throw new NotFoundException("producer not found");
         soundProducerRepository.deleteSoundProducerById(id);
     }
 
     @Override
     @Transactional
     public void deleteProducerByNickName(String nickName) {
+        if(!soundProducerRepository.existsSoundProducerByPersonInfoNickname(nickName))
+            throw new NotFoundException("producer not found");
         soundProducerRepository.deleteSoundProducerByPersonInfoNickname(nickName);
     }
 
     @Override
     @Transactional
     public SoundProducerResDto updateProducer(SoundProducerReqDto soundProducerReqDto) {
+        if(!soundProducerRepository.existsSoundProducerByPersonInfoNickname(soundProducerReqDto.getNickname()))
+            throw new NotFoundException("producer not found");
         SoundProducer producerToUpdate = soundProducerRepository.getSoundProducerByPersonInfoNickname(soundProducerReqDto.getNickname()).get();
         soundProducerDtoMapper.updateProducerInfo(producerToUpdate,soundProducerReqDto);
         soundProducerRepository.save(producerToUpdate);
@@ -52,11 +62,15 @@ public class SoundProducerServiceImpl implements SoundProducerService {
 
     @Override
     public SoundProducerResDto getProducerById(Integer id) {
+        if(!soundProducerRepository.existsSoundProducerById(id))
+            throw new NotFoundException("producer not found");
         return soundProducerDtoMapper.producerToProducerResDto(soundProducerRepository.getSoundProducerById(id).get());
     }
 
     @Override
     public SoundProducerResDto getProducerByNickname(String nickName) {
+        if(!soundProducerRepository.existsSoundProducerByPersonInfoNickname(nickName))
+            throw new NotFoundException("producer not found");
         return soundProducerDtoMapper.producerToProducerResDto(soundProducerRepository.getSoundProducerByPersonInfoNickname(nickName).get());
     }
 }
